requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.0
SQLAlchemy==2.0.33
python-dotenv==1.1.0
requests==2.32.3
Jinja2==3.1.3
itsdangerous==2.2.0
MarkupSafe==3.1.3
blinker==1.9.0
click==8.1.7
certifi==2025.4.25
idna==3.10
charset-normalizer==3.4.2
urllib3==2.2.0 